Add Terminal keybindings for Insert key.
Also add keybindings for F19 as a stand-in key for Insert, because Mac OS X
doesn't seem to support a true Insert key. Using Help instead of Insert
didn't work, because the OS app framework eats the keydown event. I wasn't
able to find any other free keycodes that don't appear on my keyboard.

Note that you'll need to use KeyRemap4MacBook to bind the Fn key (the key
where the Insert key should be) to F19 to make this all work.