Initial Ansible implementation
Got Ansible started with a playbook and a few roles.

Created a custom dynamic inventory script to make it easy to determine
what computer we're running on (by MAC address). This will let us use
a simple config file to decide how to configure each computer, and we
won't need to specify anything system-specific at the command line.

Note that we need at least Ansible 2.0 to run the osx_defaults module.