Add key bindings for function keys in Terminal
Includes all combinations of Shift, Option, and Control modifiers.

Note that this is in the shell script, not the Ansible playbook. That's
because I don't have the keybinding code moved over to Ansible yet. I
mainly just wanted to get all these bindings documented.

Also note that defaults+ (which this script uses) is currently broken on
(at least) macOS Sierra.