Install Macs Fan Control for newer MacBooks
I prefer smcFanControl, as it's easier to configure different fan speeds.
But the current version doesn't support newer MacBook Pro models.
We assume that if it has a Touch Bar, it's not supported by smcFanControl.
This is probably a pretty safe assumption, as it's likely the T2 chip that
makes the difference.